kestra/scripts/requirements.txt
kestra
google.generativeai
httpx
supabase